chore: minor changes for storage engine(#164)
- renamed: TSSPFileReader -> tsspFileReader.
- renamed: TableReader -> TSSPFileReader.
- renamed: engine/immutable/tssp_file.go -> engine/immutable/tssp_file_reader.go.
- renamed: engine/immutable/tssp_reader.go -> engine/immutable/tssp_file.go.
- renamed: engine/immutable/tssp_reader_test.go -> engine/immutable/tssp_file_reader_test.go.
- move TSSPFileReader from engine/immutable/reader.go to engine/immutable/tssp_file.go.
- move TablesStore from engine/immutable/tssp_reader.go to engine/immutable/mms_tables.go.
- GetMsBuilder+AllocMsBuilder = NewMsBuilder, GetMsBuilder create a new MsBuilder partially,
  should not be named as Getxxx.
- fileWriter -> tsspFileWriter.
- refactor for tsspfile and tsspfilereader, and make tsspfilereader as part of tsspfile.
- split writer.go:
  - move pre aggregation related codes to pre_aggregation.go.
  - move tssp file meta related codes to meta.go, including Segment, ColumnMeta, ChunkMeta,
    MetaIndex etc.
- create encoding package:
  - renamed: engine/immutable/{ => encoding}/bool.go.
  - renamed: engine/immutable/{ => encoding}/encoding.go.
  - renamed: engine/immutable/{ => encoding}/encoding_test.go.
  - renamed: engine/immutable/{ => encoding}/float.go.
  - renamed: engine/immutable/{ => encoding}/int.go.
  - renamed: engine/immutable/{ => encoding}/string.go.
  - renamed: engine/immutable/{ => encoding}/timestamp.go.
  - renamed: engine/immutable/{ => encoding}/timestamp_test.go.
