+++ engine/immutable/encoding/int.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package immutable
+package encoding
 
 import (
 	"encoding/binary"
