+++ engine/immutable/stream_compact_test.go
@@ -119,7 +119,7 @@ func TestFileIterator(t *testing.T) {
 	for i := 0; i < filesN; i++ {
 		ids, data := genTestData(idMinMax.min, idCount, recRows, &startValue, &tm)
 		fileName := NewTSSPFileName(store.NextSequence(), 0, 0, 0, true, &lockPath)
-		msb := AllocMsBuilder(store.path, "mst", &lockPath, conf, 10, fileName, store.Tier(), nil, 2)
+		msb := NewMsBuilder(store.path, "mst", &lockPath, conf, 10, fileName, store.Tier(), nil, 2)
 		write(ids, data, msb)
 
 		for j, id := range ids {
@@ -332,7 +332,7 @@ func TestMmsTables_LevelCompact_20ID10Segment_SegLimit(t *testing.T) {
 	for i := 0; i < filesN; i++ {
 		ids, data := genTestData(idMinMax.min, idCount, recRows, &startValue, &tm)
 		fileName := NewTSSPFileName(store.NextSequence(), 0, 0, 0, true, &lockPath)
-		msb := AllocMsBuilder(store.path, "mst", &lockPath, conf, 10, fileName, store.Tier(), nil, 2)
+		msb := NewMsBuilder(store.path, "mst", &lockPath, conf, 10, fileName, store.Tier(), nil, 2)
 		write(ids, data, msb)
 
 		for j, id := range ids {
@@ -606,7 +606,7 @@ func TestFileSizeExceedLimit(t *testing.T) {
 	oldIds := make([]uint64, 0, filesN)
 	for i := 0; i < filesN; i++ {
 		fileName := NewTSSPFileName(store.NextSequence(), 0, 0, 0, true, &lockPath)
-		msb := AllocMsBuilder(store.path, "mst", &lockPath, conf, 1, fileName, store.Tier(), nil, size)
+		msb := NewMsBuilder(store.path, "mst", &lockPath, conf, 1, fileName, store.Tier(), nil, size)
 		rec := genIntColumn(&startTime, rows, 2)
 		if err := msb.WriteData(id, rec); err != nil {
 			t.Fatal(err)
@@ -706,7 +706,7 @@ func TestFileSizeExceedLimit1(t *testing.T) {
 	oldIds := make([]uint64, 0, filesN)
 	for i := 0; i < filesN; i++ {
 		fileName := NewTSSPFileName(store.NextSequence(), 0, 0, 0, true, &lockPath)
-		msb := AllocMsBuilder(store.path, "mst", &lockPath, conf, 1, fileName, store.Tier(), nil, 1024)
+		msb := NewMsBuilder(store.path, "mst", &lockPath, conf, 1, fileName, store.Tier(), nil, 1024)
 		rec := genIntColumn(&startTime, rows, 2)
 		if err := msb.WriteData(id, rec); err != nil {
 			t.Fatal(err)
